refactor: TodoList에서 사용하는 hooks를 Context로부터 제공받도록 수정
- 기존에 App -> TodoOutput -> TodoList로 hooks를 넘겨받는 구조를 보다 간단하게 만들기 위해 Context사용
- Context로 활용하기위해 updateStatus, deleteTodo함수를 App으로 끌어올림
