+++ src/components/TodoContext.jsx
@@ -1,0 +1,5 @@
+import React, { createContext } from "react";
+
+const TodoContext = createContext();
+
+export default TodoContext;
