+++ src/components/TodoList.jsx
@@ -1,10 +1,11 @@
-import React from "react";
+import React, { useContext } from "react";
 import styled from "styled-components";
 import Li from "../atomicComponents/Li.jsx";
 import XButton from "../atomicComponents/XButton.jsx";
+import TodoContext from "./TodoContext.jsx";
 
 const StyledLi = styled(Li)`
-  text-decoration: ${props => (props.textDeco === "done" ? "line-through" : "none")};
+  text-decoration: ${({ textDeco }) => (textDeco === "done" ? "line-through" : "none")};
   text-decoration-color: black;
 `;
 
@@ -12,10 +13,11 @@ const Ul = styled.ul`
   padding-left: 2rem;
 `;
 
-const TodoList = ({ contents, updateStatus, deleteTodo }) => {
+const TodoList = () => {
+  const { todos, updateStatus, deleteTodo } = useContext(TodoContext);
   return (
     <Ul>
-      {contents.map(v => (
+      {todos.map(v => (
         <StyledLi textDeco={v.status} key={v.id} data-id={v.id} onClick={updateStatus}>
           {v.title}
           <XButton onClick={deleteTodo} />
